Remove dead PaddingInfoAttr _and_ use structs for tuning parameters (#728)
* Remove dead PaddingInfoAttr

Since we now use bounds inference to determine when bounds checks are
needed, and since the bugs that prevented us from using index diffs in
padding kernels are long gone, there are no uses of the
PaddingInfoAttr in any of the gemm implementations ... but we still
pass it in.

Remove the attribute, remove the code setting it, and rework any
relevant tests to account for the loss of the information about
padding being applied.

This is also setup for a miopen.gemm operation.

* Store tuning parameters in structs (#731)

* Adding tuning params structs, doesn't build, rebase over this

* Working transition, now we fix up the tests

* Fix tests to match new scheme

Also, since blockSize and gridSize are 32-bit, actually make them
32-bit in our code.

* Use the getAttr() method

+++ mlir/include/mlir/Dialect/MIOpen/Passes.h
@@ -64,8 +64,8 @@ std::unique_ptr<Pass> createMIOpenLoopsToCfPass();
 
 /// Create a pass to affix tuning parameters to gridwise gemm ops.
 std::unique_ptr<Pass>
-createAffixTuningParametersPass(int64_t blockSizeOverride = 0,
-                                int64_t gridSizeOverride = 0,
+createAffixTuningParametersPass(uint32_t blockSizeOverride = 0,
+                                uint32_t gridSizeOverride = 0,
                                 bool fallBackNoConfig = false);
 
 #define GEN_PASS_REGISTRATION
